+++ http/file_server.ts
@@ -13,7 +13,8 @@ import { parse } from "../flags/mod.ts";
 import { assert } from "../_util/asserts.ts";
 import { red } from "../fmt/colors.ts";
 import { compareEtag, createCommonResponse } from "./util.ts";
-import { DigestAlgorithm, toHashString } from "../crypto/mod.ts";
+import { DigestAlgorithm } from "../crypto/crypto.ts";
+import { toHashString } from "../crypto/to_hash_string.ts";
 import { createHash } from "../crypto/_util.ts";
 import { VERSION } from "../version.ts";
 interface EntryInfo {
